+++ rocksdb/plugin/zenfs/fs/fs_zenfs.cc
@@ -1237,7 +1237,7 @@ void ZenFS::SetResetScheme(uint32_t r,uint32_t partial_reset_scheme,uint64_t T,u
 
 }
 
-uint64_t ZenFS::GetMaxInvalidateCompactionScore(std::vector<uint64_t>& file_candidates,uint64_t * candidate_size) {
+double ZenFS::GetMaxInvalidateCompactionScore(std::vector<uint64_t>& file_candidates,uint64_t * candidate_size) {
   // return file_candidates.size();
   uint64_t ret = 100* zbd_->GetMaxSameZoneScore(file_candidates);
   return ret;
